requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-filter>=24.0,<25.0
